+++ src/main/java/timelessodyssey/states/MainMenuState.java
@@ -28,7 +28,7 @@ public class MainMenuState extends State<MainMenu> {
     }
 
     @Override
-    public void step(Game game, GUI gui, long time) throws IOException, URISyntaxException, FontFormatException {
+    public void step(Game game, GUI gui, double time) throws IOException, URISyntaxException, FontFormatException {
         GUI.Action action = gui.getNextAction();
         gui.clearAction();
         controller.step(game, action, time);
