+++ src/test/java/timelessodyssey/states/GameStateTest.java
@@ -37,7 +37,7 @@ public class GameStateTest {
 
     @Test
     public void step() throws IOException, URISyntaxException, FontFormatException {
-        long time = 0;
+        double time = 0;
         Mockito.when(gui.getNextAction()).thenReturn(GUI.Action.NONE);
         this.state = new GameState(model){
             @Override
